Fix code style in net.

+++ src/net/download.cpp
@@ -143,7 +143,7 @@ unsigned long Download::fadler32(FILE *const file)
     rewind(file);
 
     // Calculate Adler-32 checksum
-    char *const buffer = new char[fileSize];
+    char *const buffer = new char[static_cast<size_t>(fileSize)];
     const uInt read = static_cast<uInt>(fread(buffer, 1, fileSize, file));
     unsigned long adler = adler32(0L, Z_NULL, 0);
     adler = adler32(static_cast<uInt>(adler),
